requirements.txt
streamlit>=1.40.0
openai>=1.50.0
plotly>=5.24.0
pandas>=2.2.3
numpy>=2.1.0
fpdf2>=2.8.0
python-dotenv>=1.0.1
tiktoken>=0.8.0
pydantic>=2.9.0
